test<StringSplitterTest>: 테스트 이름 변경 및 테스트 숫자 변경
- 숫자 토큰으로 변경하는 객체로 이름 변경
- 숫자를 변경해 음수 테스트 추가
- 숫자를 변경해 소수 테스트 추가
- 숫자를 변경해 자릿수가 많은 숫자 테스트 추가
